requirements.txt
Django==2.2.7
django-model-utils==3.2.0
mysociety-django-popolo==0.1.0
Pillow==6.2.1
django-registration==3.0.1
celery==4.3.0
pytz==2019.3
django-extensions==2.2.5
django-nose==1.4.6
nose==1.3.7
loremipsum==1.0.5
django-autoslug==1.9.6
django-test-without-migrations==0.6
PyYAML==5.2